fix(canisters): harden production state machines and upgrade safety

+++ canisters/disburser/src/scheduler/debug_entrypoints.rs
@@ -11,7 +11,7 @@ pub async fn debug_rescue_tick_impl() {
 }
 
 pub async fn debug_execute_payout_plan_impl() -> bool {
-    let now_nanos = ic_cdk::api::time() as u64;
+    let now_nanos = ic_cdk::api::time();
     let now_secs = now_nanos / 1_000_000_000;
     let cfg = state::with_state(|st| st.config.clone());
     let ledger = IcrcLedgerCanister::new(cfg.ledger_canister_id);
@@ -19,7 +19,7 @@ pub async fn debug_execute_payout_plan_impl() -> bool {
 }
 
 pub async fn debug_build_payout_plan_impl() -> bool {
-    let now_nanos = ic_cdk::api::time() as u64;
+    let now_nanos = ic_cdk::api::time();
     let cfg = state::with_state(|st| st.config.clone());
     let ledger = IcrcLedgerCanister::new(cfg.ledger_canister_id);
